Implement endpoint configuration
